[refactor] Store models with own id

+++ app/src/main/java/com/example/dmitry/handheld_dictionary/ui/activity/OneGroupWordListActivity.java
@@ -20,7 +20,7 @@ public class OneGroupWordListActivity extends BaseActivity {
 
         final Intent intent = getIntent();
         if (savedInstanceState == null && intent.hasExtra(EXTRA_GROUP_ID)) {
-            final Integer groupId = intent.getIntExtra(EXTRA_GROUP_ID, -1);
+            final Long groupId = intent.getLongExtra(EXTRA_GROUP_ID, -1);
             final Fragment fragment = OneGroupWordListFragment.newInstance(groupId);
             FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
             ft.replace(R.id.fragment_one_group_word_list, fragment).commit();
